Backplane enablement (#151)
* Add debug

* Try removing both possible project prefixes

* Revert "Try removing both possible project prefixes"

This reverts commit dae81d96dc5d168bfb094b6847b0bb3e5cfaf98a.

* Hack to use backplane instead of release

* Parameterize release name

* Use a different tag midfix based on product

* More parameterization

* Fix whitespace lunacy

* Correctly identify incoming parameters

* Don't add instrumented manifest any more

* Remove debug

+++ modules/pipeline-manifest/bin/decorate-manifest.sh
@@ -77,6 +77,9 @@ done
 echo Creating $shad_filename file
 jq -s '.' < manifest-sha.badjson > $shad_filename
 rm manifest-sha.badjson 2> /dev/null
-echo Creating $instrumented_filename file
-jq -s '.' < manifest-instrumented.badjson > $instrumented_filename
-rm manifest-instrumented.badjson 2> /dev/null
+
+if [[ ! -z "$instrumented_filename" ]]; then
+  echo Creating $instrumented_filename file
+  jq -s '.' < manifest-instrumented.badjson > $instrumented_filename
+  rm manifest-instrumented.badjson 2> /dev/null
+fi
